Created folders for all pages and moved all files into them, also created the css files as modules and imported into the JSX files.
