+++ src/pages/About/AboutContact.jsx
@@ -1,0 +1,5 @@
+export default function AboutContact() {
+  return (
+    <div>AboutContact</div>
+  )
+}
