+++ src/pages/About/Approach.jsx
@@ -1,0 +1,5 @@
+export default function Approach() {
+  return (
+    <div>Approach</div>
+  )
+}
