+++ src/pages/About/Mission.jsx
@@ -1,0 +1,5 @@
+export default function Mission() {
+  return (
+    <div>Mission</div>
+  )
+}
