+++ src/pages/Assessments/Assessments.jsx
@@ -1,0 +1,11 @@
+import styles from './Assessments.module.css'
+
+const Assessments = () => {
+  return (
+    <>
+      <div className={styles.container} style={{ fontSize: "3rem" }}>Assessments</div>
+    </>
+  );
+};
+
+export default Assessments;
