--- src/pages/DashPages/Assessments.jsx
@@ -1,9 +1,0 @@
-const Assessments = () => {
-  return (
-    <>
-      <div style={{ fontSize: "3rem" }}>Assessments</div>
-    </>
-  );
-};
-
-export default Assessments;
