--- src/pages/DashPages/Games.jsx
@@ -1,9 +1,0 @@
-const Games = () => {
-  return (
-    <>
-      <div style={{ fontSize: "3rem" }}>Games</div>
-    </>
-  );
-};
-
-export default Games;
