--- src/pages/DashPages/Profile.jsx
@@ -1,9 +1,0 @@
-const UserProfile = () => {
-  return (
-    <>
-      <div style={{ fontSize: "3rem" }}>Profile</div>
-    </>
-  );
-};
-
-export default UserProfile;
