--- src/pages/DashPages/Rewards.jsx
@@ -1,9 +1,0 @@
-const Rewards = () => {
-  return (
-    <>
-      <div style={{ fontSize: "3rem" }}>Rewards</div>
-    </>
-  );
-};
-
-export default Rewards;
