--- src/pages/DashPages/Settings.jsx
@@ -1,9 +1,0 @@
-const Settings = () => {
-  return (
-    <>
-      <div style={{ fontSize: "3rem" }}>Settings</div>
-    </>
-  );
-};
-
-export default Settings;
