--- src/pages/DashPages/Subjects.jsx
@@ -1,9 +1,0 @@
-const Subjects = () => {
-  return (
-    <>
-      <div style={{ fontSize: "3rem" }}>Subjects</div>
-    </>
-  );
-};
-
-export default Subjects;
