+++ src/pages/Games/Games.jsx
@@ -1,0 +1,11 @@
+import styles from './Games.module.css'
+
+const Games = () => {
+  return (
+    <>
+      <div className={styles.container} style={{ fontSize: "3rem" }}>Games</div>
+    </>
+  );
+};
+
+export default Games;
