+++ src/pages/Profile/Profile.jsx
@@ -1,0 +1,11 @@
+import styles from './Profile.module.css'
+
+const UserProfile = () => {
+  return (
+    <>
+      <div className={styles.container} style={{ fontSize: "3rem" }}>Profile</div>
+    </>
+  );
+};
+
+export default UserProfile;
