+++ src/pages/Rewards/Rewards.jsx
@@ -1,0 +1,11 @@
+import styles from './Rewards.module.css'
+
+const Rewards = () => {
+  return (
+    <>
+      <div className={styles.container} style={{ fontSize: "3rem" }}>Rewards</div>
+    </>
+  );
+};
+
+export default Rewards;
