+++ src/pages/Settings/Settings.jsx
@@ -1,0 +1,11 @@
+import styles from './Settings.module.css'
+
+const Settings = () => {
+  return (
+    <>
+      <div className={styles.container} style={{ fontSize: "3rem" }}>Settings</div>
+    </>
+  );
+};
+
+export default Settings;
