+++ src/pages/Subjects/Subjects.jsx
@@ -1,0 +1,11 @@
+import styles from './Subjects.module.css'
+
+const Subjects = () => {
+  return (
+    <>
+      <div className={styles.container} style={{ fontSize: "3rem" }}>Subjects</div>
+    </>
+  );
+};
+
+export default Subjects;
